Update extension name
