+++ docs/updates/v3_0.md
@@ -9,7 +9,7 @@ version: 3.0
 
 The Visual Studio Code (VSCode) extension for PSRule now lives side-by-side with core components of [PSRule on GitHub][1].
 
-As part of this change we are now publishing the extension as a **verified** Microsoft extension with the ID `ps-rule.code`.
+As part of this change we are now publishing the extension as a **verified** Microsoft extension with the ID `ps-rule.vscode-ps-rule`.
 
 We hope this will not only help the community to log issues and get help on the correct repository,
 but also streamline how we deliver updates in the future.
