+++ src/vscode-ps-rule/commands/configureSettings.ts
