+++ src/vscode-ps-rule/commands/createOptionsFile.ts
