+++ src/vscode-ps-rule/commands/openOptionsFile.ts
