+++ src/vscode-ps-rule/commands/walkthroughCopySnippet.ts
