+++ src/vscode-ps-rule/configuration.ts
