+++ src/vscode-ps-rule/consts.ts
