+++ src/vscode-ps-rule/docLens.ts
