+++ src/vscode-ps-rule/logger.ts
