+++ src/vscode-ps-rule/main.ts
