+++ src/vscode-ps-rule/powershell.ts
