+++ src/vscode-ps-rule/tasks.ts
