+++ src/vscode-ps-rule/test/runTest.ts
