+++ src/vscode-ps-rule/test/suite/configuration.test.ts
