+++ src/vscode-ps-rule/test/suite/extension.test.ts
