+++ src/vscode-ps-rule/test/suite/index.ts
