+++ src/vscode-ps-rule/test/suite/tasks.test.ts
