+++ src/vscode-ps-rule/utils.ts
